# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mon API Blog",
        "description": "Cette API permet de gérer un blog.",
        "contact": {
            "name": "Joséphin Sylvère",
            "email": "josephinsylvere@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/categories": {
            "get": {
                "tags": [
                    "Catégories"
                ],
                "summary": "Récupère toutes les catégories",
                "description": "Retourne une liste de toutes les catégories avec leurs posts associés.",
                "operationId": "getAllCategories",
                "responses": {
                    "200": {
                        "description": "Opération réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Category"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                }
            },
            "post": {
                "tags": [
                    "Catégories"
                ],
                "summary": "Crée une nouvelle catégorie",
                "description": "Enregistre une nouvelle catégorie avec les données fournies après validation.",
                "operationId": "createCategorie",
                "requestBody": {
                    "description": "Données de la nouvelle catégorie",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Nom de la catégorie",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Catégorie créée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Category"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                }
            }
        },
        "/api/categories/{id}": {
            "get": {
                "tags": [
                    "Catégories"
                ],
                "summary": "Récupère une catégorie par son ID",
                "description": "Retourne une catégorie et ses posts associés par son ID.",
                "operationId": "getCategorieById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la catégorie",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Opération réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Category"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                }
            },
            "put": {
                "tags": [
                    "Catégories"
                ],
                "summary": "Met à jour une catégorie",
                "description": "Met à jour une catégorie existante avec les données fournies après validation.",
                "operationId": "updateCategorie",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la catégorie à mettre à jour",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Données mises à jour de la catégorie",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Nom de la catégorie",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Catégorie mise à jour",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Category"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Catégories"
                ],
                "summary": "Supprime une catégorie",
                "description": "Supprime une catégorie et renvoie un message de succès.",
                "operationId": "deleteCategorie",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la catégorie à supprimer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Catégorie supprimée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Message de succès",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                }
            }
        },
        "/api/categories/{id}/posts": {
            "get": {
                "tags": [
                    "Catégories"
                ],
                "summary": "Récupère tous les posts d'une catégorie",
                "description": "Retourne tous les posts associés à une catégorie spécifique.",
                "operationId": "getCategoryPosts",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la catégorie",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Opération réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Post"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Category": {
                "title": "Categorie",
                "description": "Un modèle de catégorie pour le blog",
                "properties": {
                    "id": {
                        "description": "L'identifiant unique de la catégorie",
                        "type": "integer"
                    },
                    "name": {
                        "description": "Le nom de la catégorie",
                        "type": "string"
                    },
                    "slug": {
                        "description": "Le slug de la catégorie",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "La date et l'heure de création de la catégorie",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "La date et l'heure de la dernière mise à jour de la catégorie",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "Post": {
                "title": "Post",
                "description": "Un modèle de post pour le blog",
                "properties": {
                    "id": {
                        "description": "L'identifiant unique du post",
                        "type": "integer"
                    },
                    "title": {
                        "description": "Le titre du post",
                        "type": "string"
                    },
                    "content": {
                        "description": "Le contenu du post",
                        "type": "string"
                    },
                    "user_id": {
                        "description": "L'identifiant de l'utilisateur qui a créé le post",
                        "type": "integer"
                    },
                    "created_at": {
                        "description": "La date et l'heure de création du post",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "La date et l'heure de la dernière mise à jour du post",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "auth_bearer": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Entrez le jeton JWT de cette manière: Bearer <token>"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}